Fix NullReferenceException in HandleGetBranchPointsMessage()
Following 8ead96b, Method.SequencePoints and Method.BranchPoints can
never be null. However, this was still occasionally checked, often useless,
but more harmful in BasePersistance.GetSequencePointsForFunction(), where
it was assumed that a skipped method would have a null SequencePoints.
As a result, the null MethodPoint (which is null and therefore not a
SequencePoint) would be added to the return array, and would eventually
throw a NullReferenceException at "point.offset = origPoints[index].Offset;"
in MessageHandler.HandleGetBranchPointsMessage().